コスメの入力フォームをForm.vueに入れたよ！
