+++ src/views/Edit.vue
@@ -7,30 +7,14 @@
         <h2>{{ type }}の編集画面</h2>
         <ul class="list">
           <li v-for="item in list" :key="item.id" class="item"> {{ item.name }}
-            <div class="brand">{{ item.id }}</div>
+           <div class="brand">{{ item.id }}</div>
             <div class="brand">{{ item.brand }}</div>
             <div class="color">{{ item.color }}</div>
             <div class="image">{{ item.theme }}</div>
           </li>
         </ul>
         <button v-on:click="editAddButtonClicked()" >{{ cosmeAddFormButtonValue }}</button>
-        <div v-if="isShow">
-        <div>コスメの分野:{{type}}</div>
-        <div>コスメのブランド:<input v-model="cosmeBrandText" type="text" name="brand"></div>
-        <div>コスメの名前:<input v-model="cosmeNameText" type="text" name="name" ></div>
-        <div>コスメの色味:<input v-model="cosmeColorText" type="text" name="color" ></div>
-        <div>コスメのテーマ:
-          <input v-model="cosmeThemeCheckbox" value="spring" type="checkbox">
-          <label>春</label>
-          <input v-model="cosmeThemeCheckbox" value="summer" type="checkbox">
-          <label>夏</label>
-          <input v-model="cosmeThemeCheckbox" value="autumn" type="checkbox">
-          <label>秋</label>
-          <input v-model="cosmeThemeCheckbox" value="winter" type="checkbox">
-          <label>冬</label>
-        </div>
-        <button v-on:click="saveForm(type)">コスメを登録</button>
-        </div>
+        <Inputform v-if="isShow" :type="type" />
       </main>
     </div>
   </div>
@@ -39,12 +23,14 @@
 <script>
 import Header from '@/components/Header.vue'
 import Sidebar from '@/components/Sidebar.vue'
+import Inputform from '@/components/Form.vue'
 
 export default {
   name: 'edit',
   components: {
     Header,
-    Sidebar
+    Sidebar,
+    Inputform
   },
   props: {
     type: {
@@ -69,23 +55,6 @@ export default {
       }else{
         this.cosmeAddFormButtonValue = 'コスメを追加'
       }
-    },
-    saveForm(type){
-      const item = {
-        type,
-        info: {
-          brand: this.cosmeBrandText,
-          name: this.cosmeNameText,
-          color: this.cosmeColorText,
-          theme: this.cosmeThemeCheckbox
-        }
-      }
-      this.$store.dispatch('userData/registerCosmeInformation', item)
-      this.$store.dispatch('userData/loadMain')
-      this.cosmeBrandText = ''
-      this.cosmeNameText = ''
-      this.cosmeColorText = ''
-      this.cosmeThemeCheckbox = []
     }
   },
   computed: {
